Add a base presenter that will work as a default
