+++ config/i18n-tasks.yml
@@ -117,6 +117,7 @@ ignore_unused:
   - decidim.features.accountability.settings.*
   - decidim.accountability.models.project.valid_statuses.*
   - decidim.accountability.admin_log.*
+  - decidim.log.base_presenter.*
   - decidim.debates.debates.filters.*
   - decidim.events.*
